implement renderer::initialize

+++ particle/ParticleEmitter.cpp
@@ -1,11 +1,11 @@
 #include "ParticleEmitter.h"
 #include "ParticleSOA.h"
 
-SimpleEmitter::SimpleEmitter(double timeSpan) : m_TimeSpan(timeSpan)
+SimpleEmitter::SimpleEmitter(float timeSpan) : m_TimeSpan(timeSpan)
 {
 }
 
-std::vector<ParticleInitialValue> SimpleEmitter::Generates(double dt)
+std::vector<ParticleInitialValue> SimpleEmitter::Generates(float dt)
 {
 	//m_Counter += dt;
 	//if (m_Counter >= m_TimeSpan)
